some optimizations

+++ install.sh
@@ -1,41 +1,41 @@
 #!/bin/bash -e
 
-if [ ! -f /usr/local/bin/pihole ]; then
-	echo "$(date) - Installing Pi-hole..."
-	curl -sSLN https://install.pi-hole.net | sudo bash
+if [ "$1" != "un" ]; then
+	if [ ! -f /usr/local/bin/pihole ]; then
+		echo "$(date) - Installing Pi-hole..."
+		curl -sSL https://install.pi-hole.net | sudo bash
+	fi
+
+	echo "$(date) - Verifying Dependencies..."
+
+	apt-get remove -y speedtest-cli
+	if [ ! -f /usr/bin/speedtest ]; then
+		echo "$(date) - Adding speedtest source..."
+		# https://www.speedtest.net/apps/cli
+		curl -sSLN https://packagecloud.io/install/repositories/ookla/speedtest-cli/script.deb.sh | sudo bash
+	fi
+	PHP_VERSION=$(php -v | tac | tail -n 1 | cut -d " " -f 2 | cut -c 1-3)
+	apt-get install -y sqlite3 $PHP_VERSION-sqlite3 jq speedtest
+
+	echo "$(date) - Downloading Latest Speedtest Mod..."
+
+	cd /var/www/html
+	rm -rf new_admin
+	git clone --depth=1 https://github.com/ipitio/AdminLTE new_admin
+	cd new_admin
+	git fetch --tags -q
+	latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+	#git checkout $latestTag
+
+	cd /opt/
+	rm -rf new_pihole
+	git clone --depth=1 -b ipitio https://github.com/ipitio/pi-hole new_pihole
+	cd new_pihole
+	git fetch --tags -q
+	latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
+	#git checkout $latestTag
+	chmod +x advanced/Scripts/webpage.sh
 fi
-
-echo "$(date) - Verifying Dependencies..."
-
-PHP_VERSION=$(php -v | tac | tail -n 1 | cut -d " " -f 2 | cut -c 1-3)
-apt-get install sqlite3 $PHP_VERSION-sqlite3 jq -y
-apt-get remove speedtest-cli -y
-
-if [ ! -f /usr/bin/speedtest ]; then
-	echo "$(date) - Installing speedtest..."
-	# https://www.speedtest.net/apps/cli
-	curl -sSLN https://packagecloud.io/install/repositories/ookla/speedtest-cli/script.deb.sh | sudo bash
-	sudo apt-get install speedtest -y
-fi
-
-echo "$(date) - Downloading Latest Speedtest Mod..."
-
-cd /var/www/html
-rm -rf new_admin
-git clone --depth=1 https://github.com/ipitio/AdminLTE new_admin
-cd new_admin
-git fetch --tags -q
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
-#git checkout $latestTag
-
-cd /opt/
-rm -rf new_pihole
-git clone --depth=1 -b ipitio https://github.com/ipitio/pi-hole new_pihole
-cd new_pihole
-git fetch --tags -q
-latestTag=$(git describe --tags `git rev-list --tags --max-count=1`)
-#git checkout $latestTag
-chmod +x advanced/Scripts/webpage.sh
 
 db=$([ "$1" == "up" ] && echo "$3" || [ "$1" == "un" ] && echo "$2" || echo "$1")
 curl -sSLN https://github.com/ipitio/pihole-speedtest/raw/ipitio/uninstall.sh | sudo bash -s -- $db
@@ -47,9 +47,6 @@ fi
 
 if [ "$1" == "up" ]; then
 	echo "$(date) - Updating Pi-hole..."
-	cd /opt/pihole
-	git reset --hard origin/master
-	git checkout master
 	cd /var/www/html/admin
 	git reset --hard origin/master
 	git checkout master
@@ -72,7 +69,7 @@ rm -rf org_admin
 mv admin org_admin
 cp -r new_admin mod_admin
 mv new_admin admin
-cd -
+cd - > /dev/null
 cp pihole/webpage.sh.mod pihole/webpage.sh
 
 if [ ! -f /etc/pihole/speedtest.db ] || [ "$db" == "db" ]; then
